casting to number

+++ src/server/queries/projects.ts
@@ -25,7 +25,7 @@ export const findProjectId = projectName => {
 export const findProjects = () => {
   return {
     // tslint:disable-next-line:max-line-length
-    text: `SELECT project_name as "projectName", p.id, count(DISTINCT(s.name)) as "scenarioCount", count(i.id) as "itemCount", MAX(i.start_time) as "latestRun"
+    text: `SELECT project_name as "projectName", p.id, count(DISTINCT(s.name))::int as "scenarioCount", count(i.id)::int as "itemCount", MAX(i.start_time) as "latestRun"
     FROM jtl.projects as p
     LEFT JOIN jtl.scenario as s ON s.project_id = p.id
     LEFT JOIN jtl.items as i ON i.scenario_id = s.id
